Target .NET Standard 1.0.
This commit makes changes which (locally, at least) allow the targeting of
.NET Standard 1.0. Summarised, those changes are:

o The main assembly (but NOT the testing assembly) now uses the new
  .csproj format to enable builds for .NET 4.5, 4.6, and Standard 1.0.

o A number of shims are introduced to provide the objects/methods that
  .NET Standard 1.0 doesn't provide. These are all internal, and so have
  no bearing on the public API.

o Information in the [AssemblyInfo.cs] file is commented out (but not
  removed) because the file is now autogenerated by the build tools from
  information in the .csproj file.

The old branch named 'dotnet-std' has been removed, so tough luck if you
were using it (you shouldn't have been, it didn't work). The old branch
used 'project.json', whereas this one uses .csproj, which seems to be the
preferred way of doing things now.

See also: issue #2.