Convert test assembly to .NET Core.
MSTest works with .NET Core, but not .NET Standard. However, they should
be identical, so it doesn't matter too much. Tests for Standard and Core
is the ideal situation, but we're not in that situation and there's lots
of converting to be done if we're to use xUnit.

Tests for [BinaryFormatter] are omitted because it isn't in .NET Core.

Other minor (i.e. no behaviour-changing) adaptions are also made.