Reinstate [BinaryFormatter] tests.
The tests were removed in the previous commit because they don't
work under .NET Core/.NET Standard. They are now added back, but
conditionally compiled in for .NET Framework targets.

Ideally [BinaryFormatter] could be removed (and with it the shim
code that was added to the main library), but that is a breaking
change and means we'd have to bump the major version number.